+++ config/deploy.rb
@@ -3,7 +3,7 @@ require "bundler/capistrano"
 
 server "176.58.98.122", :web, :app, :db, primary: true
 
-set :application, "Pubhub"
+set :application, "MundoTornado"
 set :user, "deployer"
 set :deploy_to, "/home/#{user}/apps/#{application}"
 set :deploy_via, :remote_cache