+++ config/environments/production.rb
@@ -1,4 +1,4 @@
-Pubhub::Application.configure do
+MundoTornado::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
   # Code is not reloaded between requests