+++ config/environments/test.rb
@@ -1,4 +1,4 @@
-Pubhub::Application.configure do
+MundoTornado::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
   # The test environment is used exclusively to run your application's