+++ config/initializers/secret_token.rb
@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Pubhub::Application.config.secret_token = 'aabcdc11f64fdb3f8583c068fdf9e929fc540ad40d53c436e98e7c317a3ea4e5e156ef76ca874788e74b9421ef44b06729b8ce276421f02c4019d8d16a9534f7'
+MundoTornado::Application.config.secret_token = 'aabcdc11f64fdb3f8583c068fdf9e929fc540ad40d53c436e98e7c317a3ea4e5e156ef76ca874788e74b9421ef44b06729b8ce276421f02c4019d8d16a9534f7'