+++ config/initializers/session_store.rb
@@ -1,8 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-Pubhub::Application.config.session_store :cookie_store, key: '_Pubhub_session'
+MundoTornado::Application.config.session_store :cookie_store, key: '_MundoTornado_Session', domain: :all
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
 # (create the session table with "rails generate session_migration")
-# Pubhub::Application.config.session_store :active_record_store
+# MundoTornado::Application.config.session_store :active_record_store