+++ config/unicorn.rb
@@ -1,4 +1,4 @@
-root = "/home/deployer/apps/Pubhub/current"
+root = "/home/deployer/apps/MundoTornado/current"
 working_directory root
 pid "#{root}/tmp/pids/unicorn.pid"
 stderr_path "#{root}/log/unicorn.log"